Improve how `RemoteGroupAsyncMock` mocks a `RemoteGroup`.
Instead of making each `Remote` a free form `AsyncMock`, create a spec based on the component interface.
This means, trying to assess a member that is not part of the CSC interface raises an `AttributeError` exception, which is usefull to catch interface changes, like topics that are renamed and such.

It also adds functionality to catch changes in topic payloads.
For commands, create methods that check command call payloads and raise exception if a topic attribute is not part of the command definition.
For events and telemetry, add a method to create `SimpleNamespace` instances from the topics structure.