In ``auxtel/atcs.py``, update  ``ATCS.wait_for_inposition`` to improve reporting of timeout failures.
Instead of appending coroutines to the `tasks` list, use ``asyncio.create_task`` and give names to each of the tasks.
Then, instead of simply gathering the tasks, which leads to uncomprehensive  tracebacks when tasks timeouts, capture any exception and reprocess the error messages re-raising them as `RuntimeError` with a more comprehensive message.