Prepare _wait_hard_point_test_ok method in MTCS for parallelization
Update the wait method for hardpoint tests to be ok in order to make it
compatible with concurrent executions.

Using next will pop the data from the queue so it might happen that one
execution will “steal” the event for the other and vice versa. This
commit changes next to aget and adds a sleep to avoid that.